requirements.txt
matplotlib==3.8.0
openai==0.28.0
python-dotenv==1.0.1
python-telegram-bot==13.15
numpy==1.26.4
packaging==24.1
pillow==10.3.0
pyparsing==3.1.2
cycler==0.12.1
python-dateutil==2.9.0
kiwisolver==1.4.5
aiohttp==3.9.5
requests==2.32.3
cachetools==4.2.2
pytz==2024.1
APScheduler==3.6.3
tornado==6.1
importlib-resources==6.4.0
unittest2==1.1.0